requirements.txt
# Alexandria Cover Designer — Python Dependencies

# Image processing
Pillow>=10.0
opencv-python-headless>=4.8
opencv-python>=4.8; platform_system == "Darwin"
numpy>=1.24
scipy>=1.11

# AI image generation
replicate>=0.25
httpx>=0.25

# PDF handling
reportlab>=4.0
pypdf>=3.15
pikepdf>=10.0.0
PyMuPDF>=1.24.0

# Configuration
python-dotenv>=1.0

# SVG rendering (for cover compositor — requires libcairo on system)
cairosvg>=2.7

# Utilities
tqdm>=4.65
pathlib2>=2.3; python_version < "3.6"

# Google Drive sync
google-api-python-client>=2.100
google-auth-oauthlib>=1.1
google-auth-httplib2>=0.1

# Testing
pytest>=7.4
